pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
description = ""
authors = ["Dennis Robertson <dennis.m.robertson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
llama-cpp-python = "^0.3.2"
huggingface-hub = "^0.26.2"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
psycopg2 = "^2.9.10"
scikit-learn = "^1.5.2"
webcolors = "^24.11.1"
boto3 = "^1.35.70"
rapidfuzz = "^3.10.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
